refactor: Merge MainController to ArticleController
