+++ src/main/resources/templates/article-edit.html
@@ -68,8 +68,8 @@
             <div class="row">
                 <div class="col-lg-1 col-md-1"></div>
                 <div class="col-lg-10 col-md-10 col-sm-10">
-                    <input id="article-contents" th:value="${article} ? ${article.contents} : ''" name="contents" type="hidden">
-                    <div id="editSection"></div>
+                    <input id="article-contents" name="contents" type="hidden">
+                    <div id="editSection" th:text="${article} ? ${article.contents} : ''"></div>
                 </div>
                 <div class="col-lg-1 col-md-1">
                     <div class="btn-group btn-group-vertical">
